requirements.txt
llama-index
datasets
pinecone-client
openai
transformers
llama_hub
"unstructured[pdf]"